feat: refactor build tooling to use shared Rollup configuration and add Vitest support
