+++ packages/api-base/package.json
@@ -40,16 +40,11 @@
     "ky": "^1.10.0"
   },
   "devDependencies": {
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
     "@hcc/typescript-config": "workspace:*",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^28.0.6",
-    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@hcc/rollup-config": "workspace:*",
     "@types/react": "^19.1.12",
     "@types/react-dom": "^19.1.9",
     "rollup": "^4.50.1",
-    "rollup-plugin-dts": "^6.2.3",
     "typescript": "^5.9.2"
   }
 }
