+++ packages/ui/package.json
@@ -43,17 +43,10 @@
   },
   "devDependencies": {
     "@hcc/typescript-config": "workspace:*",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^28.0.6",
-    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@hcc/rollup-config": "workspace:*",
     "@types/react": "^19.1.12",
     "@types/react-dom": "^19.1.9",
     "rollup": "^4.50.1",
-    "rollup-plugin-dts": "^6.2.3",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-preserve-directives": "^0.4.0",
     "typescript": "^5.9.2"
   },
   "peerDependencies": {
